+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Notable changes to this project.
 - added 'liveAuroc' and `validationAuroc` to `get_staking_leaderboard`
 - added 'liveAuroc' and `validationAuroc` to `get_user_activities`
 - added `ruleset` to `get_competitions`
+- added `phoneNumber` and `country` to `get_user`
 
 ## [1.4.6] - 2019-03-30
 - remove total payments from leaderboard query (#13)